+++ rust-sdk/client/codama.js
@@ -1,11 +1,10 @@
 import { createFromRoot } from "codama";
-import { renderRustVisitor } from "@codama/renderers";
+import { renderVisitor } from "@codama/renderers-rust";
 import { rootNodeFromAnchor } from "@codama/nodes-from-anchor";
 import { readFileSync } from "fs";
 
 const idl = JSON.parse(readFileSync("../../target/idl/fusionamm.json", "utf8"));
-// IDL generated with anchor 0.29 does not have the metadata field so we have to add it manually
 const node = rootNodeFromAnchor(idl);
-const visitor = renderRustVisitor("./src/generated");
+const visitor = renderVisitor("./src/generated");
 const codama = createFromRoot(node);
 codama.accept(visitor);