+++ rust-sdk/client/src/gpa/token_badge.rs
@@ -15,9 +15,7 @@ use solana_client::rpc_filter::{Memcmp, RpcFilterType};
 use solana_pubkey::Pubkey;
 
 use super::fetch_decoded_program_accounts;
-use crate::{generated::shared::DecodedAccount, TokenBadge};
-
-pub const TOKEN_BADGE_DISCRIMINATOR: &[u8] = &[116, 219, 204, 229, 249, 116, 255, 150];
+use crate::{generated::shared::DecodedAccount, TokenBadge, TOKEN_BADGE_DISCRIMINATOR};
 
 #[derive(Debug, Clone)]
 pub enum TokenBadgeFilter {
@@ -39,6 +37,6 @@ pub async fn fetch_all_token_badge_with_filter(
     filters: Vec<TokenBadgeFilter>,
 ) -> Result<Vec<DecodedAccount<TokenBadge>>, Box<dyn Error>> {
     let mut filters: Vec<RpcFilterType> = filters.into_iter().map(|filter| filter.into()).collect();
-    filters.push(RpcFilterType::Memcmp(Memcmp::new_base58_encoded(0, TOKEN_BADGE_DISCRIMINATOR)));
+    filters.push(RpcFilterType::Memcmp(Memcmp::new_base58_encoded(0, &TOKEN_BADGE_DISCRIMINATOR)));
     fetch_decoded_program_accounts(rpc, filters).await
 }