+++ rust-sdk/fusionamm-sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @crypticdot/fusionamm-rust
 
+## 1.0.62
+
+### Patch Changes
+
+- 99f630c: Added MA price
+
+## 1.0.61
+
+### Patch Changes
+
+- 9ce4e13: Improved token-2022 support
+
+## 1.0.60
+
+### Patch Changes
+
+- 231557e: Updated the codama and generated rust client
+
 ## 1.0.59
 
 ## 1.0.58