+++ ts-sdk/tx-sender/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @crypticdot/fusionamm-tx-sender
 
+## 1.0.11
+
+### Patch Changes
+
+- 9ce4e13: Improved token-2022 support
+
 ## 1.0.10
 
 ### Patch Changes